feat(service): service with required game features
Within the service, the logic of the Depth-First Search algorithm (DFS) is implemented, which is in
charge of validating the cells adjacent to the prisoner and finding the route to the exit. There is
also the functionality that saves the games and results in the database, in addition to the methods
to obtain the report information.